+++ lib/capybara/minitest/spec.rb
@@ -16,6 +16,7 @@ module Capybara
          %W[refute_#{assertion} wont_have_#{assertion}]]
       end + [%w[assert_all_of_selectors must_have_all_of_selectors],
              %w[assert_none_of_selectors must_have_none_of_selectors],
+             %w[assert_any_of_selectors must_have_any_of_selectors],
              %w[assert_style must_have_style]] +
       %w[selector xpath css].flat_map do |assertion|
         [%W[assert_matches_#{assertion} must_match_#{assertion}],